Rename parser module
