+++ tapio/parser/__init__.py
@@ -5,12 +5,12 @@ This module contains the Parser class for extracting structured content
 from HTML files saved by the crawler module.
 """
 
-from tapio.parsers.config_models import (
+from tapio.parser.config_models import (
     HtmlToMarkdownConfig,
     ParserConfigRegistry,
     SiteParserConfig,
 )
-from tapio.parsers.parser import Parser
+from tapio.parser.parser import Parser
 
 __all__ = [
     "HtmlToMarkdownConfig",
